Update packages to resolve security alerts
The versions of minimist and acorn that were installed had 'moderate
severity' security alerts raised against them on GitHub. I've audited
all the versions of packages we directly depend on and upgraded them.
According to a run of `npm audit` this has cleared all warnings. The
other changes were knock-on effects of updating all the packages.

Issue: none