⬆️ Upgrade Node to v14
Some of our dependencies have security issues and we need to upgrade
them. Some of their new versions depend on Node 14 or later, so we need
to bump our version to keep up. Node v14 is now the LTS version so we
should use that instead of v12 anyway.

Issue: Scottish-Natural-Heritage/Licensing#674